+++ tests/testthat/test-plot_experiments_matrix.R
@@ -1,3 +1,3 @@
 test_that("multiplication works", {
-  expect_error(plot_experiments_matrix(iraceResults, fileName = "/"))
+  expect_error(plot_experiments_matrix(iraceResults, filename = "/"))
 })